rename fleet.io to placement.fleet.io
