+++ apis/placement/v1alpha1/commons.go
@@ -14,9 +14,9 @@ type ClusterState string
 
 const (
 	// Unprefixed labels/annotations are reserved for end-users
-	// we will add a fleet.io to designate these labels/annotations as official fleet labels/annotations.
+	// we will add a placement.fleet.io to designate these labels/annotations as official fleet labels/annotations.
 	// See https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#label-selector-and-annotation-conventions
-	fleetPrefix = "fleet.io/"
+	fleetPrefix = "placement.fleet.io/"
 
 	// CRPTrackingLabel is the label that points to the cluster resource policy that creates a resource binding.
 	CRPTrackingLabel = fleetPrefix + "parentCRP"
