+++ apis/placement/v1alpha1/resourcesnapshot_types.go
