+++ config/crd/bases/placement.fleet.io_clusterresourcesnapshots.yaml
@@ -4,9 +4,9 @@ kind: CustomResourceDefinition
 metadata:
   annotations:
     controller-gen.kubebuilder.io/version: v0.11.4
-  name: clusterresourcesnapshots.fleet.io
+  name: clusterresourcesnapshots.placement.fleet.io
 spec:
-  group: fleet.io
+  group: placement.fleet.io
   names:
     categories:
     - fleet-workload
